add progress_status section
in the presentation sildes for the project (from moodle), the
requirement is mentioned to report on the projet progress & status:

"
progress report: Report on the progress of your project and document who
did what.
	document applications & explain approach: Document your
application and explain your approach with respect to the theoretical
framework introduced in the lecture.
	Describe your implementation and highlight what is still missing
(and why).
"